+++ src/layman/layer/qgis/qml-template.qml
@@ -1,0 +1,3 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis>
+</qgis>