+++ src/layman/layer/rest_layer_style_test.py
@@ -8,9 +8,9 @@ from test import process_client
 
 
 @pytest.mark.usefixtures('ensure_layman')
-def test_get_layer_style():
-    username = 'test_get_layer_style_user'
-    layername = 'test_get_layer_style_layer'
+def test_get_layer_style_sld():
+    username = 'test_get_layer_style_sld_user'
+    layername = 'test_get_layer_style_sld_layer'
 
     process_client.publish_layer(username,
                                  layername,
@@ -23,4 +23,28 @@ def test_get_layer_style():
     # lxml does not support importing from utf8 string
     xml_tree = ET.fromstring(bytes(rv.text, encoding='utf8'))
     assert ET.QName(xml_tree) == "{http://www.opengis.net/sld}StyledLayerDescriptor", rv.text
+    process_client.delete_layer(username, layername)
+
+
+@pytest.mark.usefixtures('ensure_layman')
+def test_get_layer_style_qml():
+    username = 'test_get_layer_style_sld_user'
+    layername = 'test_get_layer_style_sld_layer'
+    qml_file = 'sample/style/small_layer.qml'
+
+    process_client.publish_layer(username,
+                                 layername,
+                                 style_file=qml_file
+                                 )
+
+    with app.app_context():
+        rest_url = url_for('rest_layer_style.get', username=username, layername=layername)
+    rv = requests.get(rest_url)
+    assert rv.status_code == 200, rv.text
+    # lxml does not support importing from utf8 string
+    xml_el = ET.fromstring(bytes(rv.text, encoding='utf8'))
+    assert ET.QName(xml_el), rv.text
+    assert ET.QName(xml_el) == "qgis", rv.text
+    assert len(xml_el.xpath('/qgis/renderer-v2')) == 1, rv.text
+    assert xml_el.attrib, rv.text
     process_client.delete_layer(username, layername)