+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - [#154](https://github.com/jirik/layman/issues/154) During startup, [LAYMAN_OUTPUT_SRS_LIST](doc/env-settings.md#LAYMAN_OUTPUT_SRS_LIST) is ensure for all QGIS layers.
 - [#154](https://github.com/jirik/layman/issues/154) Treat attribute names in QML (also known as '[launder](https://gdal.org/drivers/vector/pg.html#layer-creation-options)').
 - [#154](https://github.com/jirik/layman/issues/154) Endpoint [GET Layer](doc/rest.md#get-layer) returns in 'style' also 'type', either 'sld' or 'qml'.
+- [#154](https://github.com/jirik/layman/issues/154) Endpoint [GET Layer Style](doc/rest.md#get-layer-style) returns 'sld' or 'qml' XML in response.
 
 ## v1.9.1
  2021-01-18